checkpoint 2. Everything with age is working and creating a user in firestore. TODO - workout duration, training focus, schedule performance, target weight.
Bug fixes - update few set to string convertions in few screens. Do not create auth when account is not created

+++ R.E.P/models/DailyProgress.swift
@@ -148,7 +148,7 @@ struct DailyProgress: Codable, Identifiable {
     }
     
     // Static helper methods
-    private static func calculateCalorieTarget(age: Int, weight: Double, height: Double, gender: String, fitnessGoal: String) -> Double {
+    private static func calculateCalorieTarget(age: Double, weight: Double, height: Double, gender: String, fitnessGoal: String) -> Double {
         // Basic BMR calculation using Mifflin-St Jeor Equation
         let bmr: Double
         if gender == "male" {
